refactor(benchmarks): Centralize metric field name registry
The shared metric field names for tracing events are now centralized
in `metrics_registry` within the `rlevo-benchmarks` crate.

This refactoring ensures that both the live TUI layer and the on-disk
recording layer consume the same source of truth for metric names.
This prevents divergence where an algorithm emits a metric name that
is recorded but not charted, or vice versa, leading to inconsistent
visualization surfaces.

CRITICAL: The constant `CANONICAL_METRICS` definition is removed from
both `tracing_layer.rs` and `log_layer.rs` and exclusively imported
from the new `metrics_registry` module.
